Address a subtle overloading of the word "internal" in comments
Within the Packed Hilbert R-Tree code, the word "internal" had two
different meanings: it can refer to an internal tree node, i.e. a
non-leaf node; or to an internal data structure, i.e. one that is
private and not exported from the Go package.

I rewrote a few comments and some documentation to use terms like
"private" and "non-exported" to cover the latter case, reserving
"internal" for non-leaf nodes.