Drop IndexDataPtr method and simplify Data method for FileWriter
Keeping both IndexData and IndexDataPtr is unnecessarily complex,
especially given that individual flat.Feature structs are pretty
lightweight, effectively containing only the two fields required for a
FlatBuffer table.

The change to the FileWriter's Data method is intended to make it look
as much like the FileReader's Data method as possible. Of course, there
are still some subtle differences, like the former returns a byte count
and the latter returns a feature count.

This change simplifies, and it also amkes FileWriter look more like
FileReader, resulting in a simpler, more parallel, design.