Attempt to give example tests access to their data on pkg.go.dev
I'm not sure yet whether this attempt will work, but its a first try.

If this doesn't work, the next idea is to use the Go embed module to
embed the test data in the flatgeobuf_test module. I have a decent
feeling about that working, because it makes the test data part of
the module itself.