Clean up some minor code flags
The GoLand IDEA currently adds warnings flags in the margin for all
unhandled errors, even in cases where the documented behavior of the
functions being called can't result in an error. This just makes
those flags go away because a *strings.Builder won't return errors,
although it might panic if it runs out of memory.