Delete TODO from example
After the change, the search query box in the example is a bounding box
around most of Gogama, ON, CA. The first search result, however, is
Russia, which is counterintuitive. This is because of the weirdness of
WGS 84 and the fact that Siberia crosses the anti-meridian, resulting
in Russia having a bounding box spanning longitudes from -180 to 180,
so that our Gogama bounding box "intersects" it.